Add files via upload
updated some of the html and css on the main page, its the best I could get it to look while still being functional. also had to remove the return to session button in order to get the last dropdown to stay on the screen. want to try and find a way to put it next to the others but not sure how to do that at the moment.